requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
autoflake==2.3.1
Automat==24.8.1
autopep8==2.3.2
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
dash==2.9.3
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distlib==0.3.9
dj-database-url==0.5.0
Django==3.2.25
django-appconf==1.1.0
django-bootstrap4==24.1
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.5.0
django-countries==7.6.1
django-cryptography==1.1
django-filter==23.5
django-guardian==2.4.0
django-js-asset==2.2.0
django-plotly-dash==2.2.2
django-ratelimit==4.1.0
django-unfold==0.44.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
dpd-static-support==0.0.5
dpd_components==0.1.0
drf-spectacular==0.26.5
et_xmlfile==2.0.0
filelock==3.18.0
firebase-admin==6.6.0
flake8==7.3.0
Flask==3.1.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==20.1.0
httplib2==0.22.0
hyperlink==21.0.0
identify==2.6.12
idna==3.10
incremental==24.7.2
inflection==0.5.1
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mccabe==0.7.0
msgpack==1.1.0
mypy==1.16.1
mypy_extensions==1.1.0
narwhals==1.29.0
nodeenv==1.9.1
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.8
plotly==6.0.0
pre_commit==4.2.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.3.1
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
service-identity==24.2.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.2
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.31.2
Werkzeug==3.1.3
whitenoise==5.3.0
zope.interface==7.2
